+++ crates/memory/src/arc/ptr.rs
@@ -3,7 +3,7 @@ use std::{
     fmt,
     hash::{Hash, Hasher},
     ops::Deref,
-    rc::Rc,
+    sync::Arc,
 };
 
 use crate::Address;
@@ -17,34 +17,27 @@ use crate::Address;
 #[macro_export]
 macro_rules! make_ptr {
     ($value:expr) => {{
-        use std::rc::Rc;
+        use std::sync::Arc;
 
-        Ptr::from(Rc::new($value) as Rc<_>)
+        Ptr::from(Arc::new($value) as Arc<_>)
     }};
 }
 
 /// An immutable pointer to a value in allocated memory
 #[derive(Debug, Default)]
-pub struct Ptr<T: ?Sized>(Rc<T>);
-
-impl<T> Ptr<T> {
-    /// Moves the value into newly allocated memory
-    pub fn new(value: T) -> Self {
-        Self(Rc::new(value))
-    }
-}
+pub struct Ptr<T: ?Sized>(Arc<T>);
 
 impl<T: ?Sized> Ptr<T> {
     /// Returns true if the two `Ptr`s point to the same allocation
     ///
-    /// See also: [std::rc::Rc::ptr_eq]
+    /// See also: [std::sync::Arc::ptr_eq]
     pub fn ptr_eq(this: &Self, other: &Self) -> bool {
-        Rc::ptr_eq(&this.0, &other.0)
+        Arc::ptr_eq(&this.0, &other.0)
     }
 
     /// Returns the address of the allocated memory
     pub fn address(this: &Self) -> Address {
-        Rc::as_ptr(&this.0).into()
+        Arc::as_ptr(&this.0).into()
     }
 }
 
@@ -55,15 +48,15 @@ impl<T: Clone> Ptr<T> {
     /// Otherwise a clone of the value will be made to ensure uniqueness before returning the
     /// reference.
     ///
-    /// See also: [std::rc::Rc::make_mut]
+    /// See also: [std::sync::Arc::make_mut]
     pub fn make_mut(this: &mut Self) -> &mut T {
-        Rc::make_mut(&mut this.0)
+        Arc::make_mut(&mut this.0)
     }
 }
 
 impl<T> From<T> for Ptr<T> {
     fn from(value: T) -> Self {
-        Self::new(value)
+        Self(Arc::from(value))
     }
 }
 
@@ -73,8 +66,8 @@ impl<T: ?Sized> From<Box<T>> for Ptr<T> {
     }
 }
 
-impl<T: ?Sized> From<Rc<T>> for Ptr<T> {
-    fn from(inner: Rc<T>) -> Self {
+impl<T: ?Sized> From<Arc<T>> for Ptr<T> {
+    fn from(inner: Arc<T>) -> Self {
         Self(inner)
     }
 }
@@ -89,41 +82,41 @@ impl<T: ?Sized> Deref for Ptr<T> {
 
 impl<T: ?Sized> Clone for Ptr<T> {
     fn clone(&self) -> Self {
-        Self(Rc::clone(&self.0))
+        Self(Arc::clone(&self.0))
     }
 }
 
 impl<T: Clone> From<&[T]> for Ptr<[T]> {
     #[inline]
     fn from(value: &[T]) -> Self {
-        Self(Rc::from(value))
+        Self(Arc::from(value))
     }
 }
 
 impl<T> From<Vec<T>> for Ptr<[T]> {
     #[inline]
     fn from(value: Vec<T>) -> Self {
-        Self(Rc::from(value))
+        Self(Arc::from(value))
     }
 }
 
 impl From<&str> for Ptr<str> {
     #[inline]
     fn from(value: &str) -> Self {
-        Self(Rc::from(value))
+        Self(Arc::from(value))
     }
 }
 
 impl From<String> for Ptr<str> {
     #[inline]
     fn from(value: String) -> Self {
-        Self(Rc::from(value))
+        Self(Arc::from(value))
     }
 }
 
 impl<T: ?Sized + PartialEq> PartialEq for Ptr<T> {
     fn eq(&self, other: &Self) -> bool {
-        Rc::eq(&self.0, &other.0)
+        Arc::eq(&self.0, &other.0)
     }
 }