+++ crates/runtime/src/types/map.rs
@@ -106,7 +106,7 @@ impl KMap {
 
     /// Provides a reference to the KMap's meta map
     ///
-    /// This is returned as a reference to the meta map's Rc to allow for cloning.
+    /// This is returned as a reference to the meta map's PtrMut to allow for cloning.
     pub fn meta_map(&self) -> Option<&PtrMut<MetaMap>> {
         self.meta.as_ref()
     }
@@ -131,8 +131,8 @@ impl KMap {
     }
 
     /// Insert an entry into the KMap's data
-    pub fn insert(&self, key: ValueKey, value: Value) {
-        self.data_mut().insert(key, value);
+    pub fn insert(&self, key: impl Into<ValueKey>, value: impl Into<Value>) {
+        self.data_mut().insert(key.into(), value.into());
     }
 
     /// Inserts a value into the meta map, initializing the meta map if it doesn't yet exist
@@ -144,18 +144,8 @@ impl KMap {
     }
 
     /// Adds a function to the KMap's data map
-    pub fn add_fn(&self, id: &str, f: impl Fn(&mut CallContext) -> Result<Value> + 'static) {
-        self.add_value(id, Value::NativeFunction(KNativeFunction::new(f)));
-    }
-
-    /// Adds a map to the KMap's data map
-    pub fn add_map(&self, id: &str, map: KMap) {
-        self.add_value(id, Value::Map(map));
-    }
-
-    /// Adds a [Value](crate::Value) to the KMap's data map
-    pub fn add_value(&self, id: &str, value: Value) {
-        self.insert(id.into(), value);
+    pub fn add_fn(&self, id: &str, f: impl KotoFunction) {
+        self.insert(id, Value::NativeFunction(KNativeFunction::new(f)));
     }
 
     /// Returns the number of entries in the KMap's data map
@@ -238,7 +228,7 @@ mod tests {
         let m = KMap::default();
 
         assert!(m.data().get("test").is_none());
-        m.add_value("test", Value::Null);
+        m.insert("test", Value::Null);
         assert!(m.data().get("test").is_some());
         assert!(matches!(m.data_mut().remove("test"), Some(Value::Null)));
         assert!(m.data().get("test").is_none());