+++ CHANGELOG.md
@@ -36,6 +36,13 @@ The Koto project adheres to
 - Objects can be compared with `null` on the LHS without having to implement 
   `KotoObject::equal` and/or `not_equal`.
 
+#### Internals
+
+- The Koto runtime is now thread-safe by default, with the previous
+  single-threaded behaviour available via the `rc` feature.
+  - The `rc` variant has slightly better performance at the cost of thread
+    safety.
+
 #### REPL
 
 - The REPL `config.koto` settings have all been moved into a `repl` sub-map.
@@ -47,6 +54,8 @@ The Koto project adheres to
 #### API
 
 - `ObjectEntryBuilder` has been replaced with macros from `koto_derive`.
+- `KMap::add_map` and `KMap::add_value` have been removed, `KMap::insert` now
+  accepts any value that implements `Into<Value>` and can be used instead.
 
 
 ## [0.12.0] 2023.10.18